--  Fixed inspector warnings. Extracted ArtistSearchTask and KeyboardUtil to own classes.
