+++ app/src/main/res/layout/item_artist_search_result.xml
@@ -8,8 +8,8 @@
 
     <ImageView
         android:id="@+id/artistImage"
-        style="@style/artist_search_result_thumb"
         tools:src="@drawable/notification_template_icon_bg"
+        android:contentDescription="@string/artist_image_content_description"
         android:layout_width="@dimen/artist_search_result_image_width"
         android:layout_height="@dimen/artist_search_result_image_height"
         android:layout_margin="@dimen/search_result_margin_top" />
