+++ app/src/main/res/values/strings.xml
@@ -2,6 +2,10 @@
     <string name="app_name">My App Portfolio</string>
     <string name="action_settings">Settings</string>
 
+    <string name="demo_artist_name">King</string>
+    <string name="artist_search_hint">Search for an artist</string>
+    <string name="artist_image_content_description">Artist image</string>
+
     <!-- menu page -->
     <string name="title">My Nanodegree Apps</string>
     <!-- app names -->
@@ -15,6 +19,4 @@
     <!-- messaging -->
     <string name="button_selected_message_prefix">This button will launch my </string>
     <string name="title_activity_artist_search">ArtistSearchActivity</string>
-    <string name="hello_world">Hello world!</string>
-
 </resources>
