+++ app/src/main/res/values/styles.xml
@@ -27,9 +27,6 @@
         <item name="android:background">@color/colorAccent</item>
     </style>
 
-    <style name="artist_search_result"></style>
-
-    <style name="artist_search_result_thumb"></style>
 
     <!--<style name="item_base_style">-->
